requirements.txt
pypdf>=4.0.0
ragas>=0.1.7
openai>=1.10.0
python-dotenv>=1.0.0
pandas>=2.1.0
tqdm>=4.66.0
jinja2>=3.1.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
numpy>=1.24.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
rank-bm25>=0.2.2
scikit-learn>=1.3.0
